Fixed all Unit Tests
