+++ events/blocks/bento-cards/bento-cards.js
@@ -1,5 +1,5 @@
 import buildMiloCarousel from '../../features/milo-carousel.js';
-import { LIBS } from '../../scripts/scripts.js';
+import { LIBS } from '../../scripts/utils.js';
 
 const { createTag } = await import(`${LIBS}/utils/utils.js`);
 const { decorateButtons } = await import(`${LIBS}/utils/decorate.js`);
