+++ events/blocks/preview-bar/preview-bar.js
@@ -1,5 +1,4 @@
-import { LIBS } from '../../scripts/scripts.js';
-import { getIcon } from '../../scripts/utils.js';
+import { LIBS, getIcon } from '../../scripts/utils.js';
 
 const { createTag, getMetadata } = await import(`${LIBS}/utils/utils.js`);
