+++ events/features/milo-carousel.js
@@ -1,8 +1,8 @@
-import { LIBS, MILO_CONFIG } from '../scripts/scripts.js';
+import { LIBS } from '../scripts/utils.js';
 
-const { createTag, MILO_EVENTS, loadStyle } = await import(`${LIBS}/utils/utils.js`);
+const { createTag, MILO_EVENTS, loadStyle, getConfig } = await import(`${LIBS}/utils/utils.js`);
 
-const { miloLibs, codeRoot } = MILO_CONFIG;
+const { miloLibs, codeRoot } = getConfig();
 const base = miloLibs || codeRoot;
 
 export const ARROW_NEXT_IMG = `<img class="next-icon" alt="Next icon" src="${base}/blocks/carousel/img/arrow.svg" height="10" width="16">`;
@@ -277,7 +277,7 @@ function handleChangingSlides(carouselElements) {
 
 export default function buildMiloCarousel(el, slides) {
   return new Promise((resolve) => {
-    loadStyle(`${MILO_CONFIG.codeRoot}/features/milo-carousel.css`, () => {
+    loadStyle(`${getConfig().codeRoot}/features/milo-carousel.css`, () => {
       const parentArea = el.closest('.fragment') || document;
       el.classList.add('carousel-plugin');
