+++ events/scripts/esp-controller.js
@@ -1,3 +1,18 @@
+export const API_CONFIG = {
+  esl: {
+    local: { host: 'http://localhost:8499' },
+    dev: { host: 'https://wcms-events-service-layer-deploy-ethos102-stage-va-9c3ecd.stage.cloud.adobe.io' },
+    stage: { host: 'https://events-service-layer-stage.adobe.io' },
+    prod: { host: 'https://events-service-layer.adobe.io' },
+  },
+  esp: {
+    local: { host: 'http://localhost:8500' },
+    dev: { host: 'https://wcms-events-service-platform-deploy-ethos102-stage-caff5f.stage.cloud.adobe.io' },
+    stage: { host: 'https://events-service-platform-stage.adobe.io' },
+    prod: { host: 'https://events-service-platform.adobe.io' },
+  },
+};
+
 export const getCaasTags = (() => {
   let cache;
   let promise;
@@ -29,23 +44,6 @@ export const getCaasTags = (() => {
     return promise;
   };
 })();
-
-export function getAPIConfig() {
-  return {
-    esl: {
-      local: { host: 'http://localhost:8499' },
-      dev: { host: 'https://wcms-events-service-layer-deploy-ethos102-stage-va-9c3ecd.stage.cloud.adobe.io' },
-      stage: { host: 'https://events-service-layer-stage.adobe.io' },
-      prod: { host: 'https://events-service-layer.adobe.io' },
-    },
-    esp: {
-      local: { host: 'http://localhost:8500' },
-      dev: { host: 'https://wcms-events-service-platform-deploy-ethos102-stage-caff5f.stage.cloud.adobe.io' },
-      stage: { host: 'https://events-service-platform-stage.adobe.io' },
-      prod: { host: 'https://events-service-platform.adobe.io' },
-    },
-  };
-}
 
 export function waitForAdobeIMS() {
   return new Promise((resolve) => {
@@ -79,7 +77,7 @@ export async function constructRequestOptions(method, body = null) {
 }
 
 export async function getEvent(eventId) {
-  const { host } = getAPIConfig().esp[window.eccEnv];
+  const { host } = API_CONFIG.esp[window.eccEnv];
   const options = await constructRequestOptions('GET');
 
   try {
@@ -99,7 +97,7 @@ export async function getEvent(eventId) {
 }
 
 export async function getEventAttendee(eventId) {
-  const { host } = getAPIConfig().esp[window.eccEnv];
+  const { host } = API_CONFIG.esp[window.eccEnv];
   const options = await constructRequestOptions('GET');
 
   try {
@@ -122,7 +120,7 @@ export async function getEventAttendee(eventId) {
 }
 
 export async function getAttendee() {
-  const { host } = getAPIConfig().esl[window.eccEnv];
+  const { host } = API_CONFIG.esl[window.eccEnv];
   const options = await constructRequestOptions('GET');
 
   try {
@@ -147,7 +145,7 @@ export async function getAttendee() {
 export async function createAttendee(attendeeData) {
   if (!attendeeData) return false;
 
-  const { host } = getAPIConfig().esl[window.eccEnv];
+  const { host } = API_CONFIG.esl[window.eccEnv];
   const raw = JSON.stringify(attendeeData);
   const options = await constructRequestOptions('POST', raw);
 
@@ -171,7 +169,7 @@ export async function addAttendeeToEvent(eventId, attendee) {
   if (!eventId || !attendee) return false;
 
   const { firstName, lastName, email } = attendee;
-  const { host } = getAPIConfig().esl[window.eccEnv];
+  const { host } = API_CONFIG.esl[window.eccEnv];
   const raw = JSON.stringify({ firstName, lastName, email });
   const options = await constructRequestOptions('POST', raw);
 
@@ -194,7 +192,7 @@ export async function addAttendeeToEvent(eventId, attendee) {
 export async function updateAttendee(attendeeData) {
   if (!attendeeData) return false;
 
-  const { host } = getAPIConfig().esl[window.eccEnv];
+  const { host } = API_CONFIG.esl[window.eccEnv];
   const raw = JSON.stringify(attendeeData);
   const options = await constructRequestOptions('PUT', raw);
 
@@ -217,7 +215,7 @@ export async function updateAttendee(attendeeData) {
 export async function deleteAttendeeFromEvent(eventId) {
   if (!eventId) return false;
 
-  const { host } = getAPIConfig().esl[window.eccEnv];
+  const { host } = API_CONFIG.esl[window.eccEnv];
   const options = await constructRequestOptions('DELETE');
 
   try {
